integrate esp-now and add game logic (got hard fault)

+++ Core/Src/freertos.c
@@ -145,7 +145,7 @@ void MX_FREERTOS_Init(void) {
 
   /* Create the queue(s) */
   /* creation of commQueue */
-  commQueueHandle = osMessageQueueNew (4, sizeof(uint16_t), &commQueue_attributes);
+  commQueueHandle = osMessageQueueNew (4, sizeof(uint32_t), &commQueue_attributes);
 
   /* creation of uartRxQueue */
   uartRxQueueHandle = osMessageQueueNew (3, sizeof(uint16_t), &uartRxQueue_attributes);
